PC用サイドバー部分の切り出し
