+++ src/components/Layouts/MSidebar.tsx
@@ -5,10 +5,7 @@ import { XMarkIcon } from '@heroicons/react/24/outline'
 import Image from 'next/image'
 import Link from 'next/link'
 import { navigation } from '@/lib/navigation/navigationData'
-
-export function classNames(...classes: string[]) {
-  return classes.filter(Boolean).join(' ')
-}
+import { classNames } from '@/lib/utils/utils'
 
 interface MSidebarProps {
   sidebarOpen: boolean
